+++ packages/vtable/src/scenegraph/scenegraph.ts
@@ -772,6 +772,9 @@ export class Scenegraph {
   }
 
   resize() {
+    // reset proxy config
+    this.proxy.resize();
+
     if (this.table.widthMode === 'adaptive' || this.table.autoFillWidth) {
       if (this.table.internalProps._widthResizedColMap.size === 0) {
         //如果没有手动调整过行高列宽 则重新计算一遍并重新分配
@@ -806,6 +809,8 @@ export class Scenegraph {
     ) {
       this.updateChartSize(this.table.rowHeaderLevelCount);
     }
+
+    this.proxy.progress();
     // this.stage.window.resize(width, height);
     this.updateNextFrame();
   }
@@ -828,12 +833,25 @@ export class Scenegraph {
       )
     } as any);
 
-    if (this.tableGroup.border) {
+    if (this.tableGroup.border && this.tableGroup.border.type === 'rect') {
       this.tableGroup.border.setAttributes({
         x: this.table.tableX - this.tableGroup.border.attribute.lineWidth / 2,
         y: this.table.tableY - this.tableGroup.border.attribute.lineWidth / 2,
         width: this.tableGroup.attribute.width + this.tableGroup.border.attribute.lineWidth,
         height: this.tableGroup.attribute.height + this.tableGroup.border.attribute.lineWidth
+      });
+    } else if (this.tableGroup.border && this.tableGroup.border.type === 'group') {
+      this.tableGroup.border.setAttributes({
+        x: this.table.tableX - this.tableGroup.border.attribute.lineWidth / 2,
+        y: this.table.tableY - this.tableGroup.border.attribute.lineWidth / 2,
+        width: this.tableGroup.attribute.width + this.tableGroup.border.attribute.lineWidth,
+        height: this.tableGroup.attribute.height + this.tableGroup.border.attribute.lineWidth
+      });
+      (this.tableGroup.border.firstChild as IRect)?.setAttributes({
+        x: this.tableGroup.border.attribute.lineWidth / 2,
+        y: this.tableGroup.border.attribute.lineWidth / 2,
+        width: this.tableGroup.attribute.width,
+        height: this.tableGroup.attribute.height
       });
     }